Move get json data to a file called ServiceData
